python2 -> python3, and inital inheritance pattern
