+++ .travis.yml
@@ -13,7 +13,7 @@ before_install:
   - docker pull shookees/kaldi-gstreamer-server
 
 before_script:
-  - pip install codeclimate-test-reporter
+  - pip3 install codeclimate-test-reporter
 
 script:
  - docker run -v `pwd`:/opt/kaldi-gstreamer-server-test shookees/kaldi-gstreamer-server bash -c "pip3 install coverage==4.3 &&
