+++ kaldigstserver/decoder/Nn2DecoderPipeline.py
@@ -1,9 +1,11 @@
 import os
-import thread
+import _thread
 
 import gi
 import logging
 from gi.repository import GObject, Gst
+
+from kaldigstserver.decoder.DecoderPipeline import DecoderPipeline
 
 gi.require_version('Gst', '1.0')
 
@@ -11,12 +13,10 @@ GObject.threads_init()
 Gst.init(None)
 
 
-logger = logging.getLogger(__name__)
-
-
-class DecoderPipeline2(object):
-    def __init__(self, conf={}):
-        logger.info("Creating decoder using conf: %s" % conf)
+class Nn2DecoderPipeline(DecoderPipeline):
+    def __init__(self, conf={}, logger=logging.getLogger(__name__)):
+        self.logger = logger
+        self.logger.info("Creating decoder using conf: %s" % conf)
         self.create_pipeline(conf)
         self.output_dir = conf.get("out-dir", None)
         if self.output_dir:
@@ -53,28 +53,28 @@ class DecoderPipeline2(object):
 
         decoder_config = conf.get("decoder", {})
         if 'nnet-mode' in decoder_config:
-            logger.info("Setting decoder property: %s = %s"
+            self.logger.info("Setting decoder property: %s = %s"
                         % ('nnet-mode', decoder_config['nnet-mode']))
             self.asr.set_property('nnet-mode', decoder_config['nnet-mode'])
             del decoder_config['nnet-mode']
 
-        for (key, val) in decoder_config.iteritems():
+        for (key, val) in decoder_config.items():
             if key != "use-threaded-decoder":
-                logger.info("Setting decoder property: %s = %s" % (key, val))
+                self.logger.info("Setting decoder property: %s = %s" % (key, val))
                 self.asr.set_property(key, val)
 
         self.appsrc.set_property("is-live", True)
         self.filesink.set_property("location", "/dev/null")
-        logger.info('Created GStreamer elements')
+        self.logger.info('Created GStreamer elements')
 
         self.pipeline = Gst.Pipeline()
         for element in [self.appsrc, self.decodebin, self.audioconvert,
                         self.audioresample, self.tee, self.queue1,
                         self.filesink, self.queue2, self.asr, self.fakesink]:
-            logger.debug("Adding %s to the pipeline" % element)
+            self.logger.debug("Adding %s to the pipeline" % element)
             self.pipeline.add(element)
 
-        logger.info('Linking GStreamer elements')
+        self.logger.info('Linking GStreamer elements')
 
         self.appsrc.link(self.decodebin)
         self.decodebin.connect('pad-added', self._connect_decoder)
@@ -101,42 +101,42 @@ class DecoderPipeline2(object):
         self.asr.connect('final-result', self._on_final_result)
         self.asr.connect('full-final-result', self._on_full_final_result)
 
-        logger.info("Setting pipeline to READY")
+        self.logger.info("Setting pipeline to READY")
         self.pipeline.set_state(Gst.State.READY)
-        logger.info("Set pipeline to READY")
+        self.logger.info("Set pipeline to READY")
 
     def _connect_decoder(self, element, pad):
-        logger.info("%s: Connecting audio decoder" % self.request_id)
+        self.logger.info("%s: Connecting audio decoder" % self.request_id)
         pad.link(self.audioconvert.get_static_pad("sink"))
-        logger.info("%s: Connected audio decoder" % self.request_id)
+        self.logger.info("%s: Connected audio decoder" % self.request_id)
 
     def _on_partial_result(self, asr, hyp):
-        logger.info("%s: Got partial result: %s"
-                    % (self.request_id, hyp.decode('utf8')))
+        self.logger.info("%s: Got partial result: %s"
+                    % (self.request_id, hyp))
         if self.result_handler:
-            self.result_handler(hyp.decode('utf8'), False)
+            self.result_handler(hyp, False)
 
     def _on_final_result(self, asr, hyp):
-        logger.info("%s: Got final result: %s"
-                    % (self.request_id, hyp.decode('utf8')))
+        self.logger.info("%s: Got final result: %s"
+                    % (self.request_id, hyp))
         if self.result_handler:
-            self.result_handler(hyp.decode('utf8'), True)
+            self.result_handler(hyp, True)
 
     def _on_full_final_result(self, asr, result_json):
-        logger.info("%s: Got full final result: %s"
-                    % (self.request_id, result_json.decode('utf8')))
+        self.logger.info("%s: Got full final result: %s"
+                    % (self.request_id, result_json))
         if self.full_result_handler:
             self.full_result_handler(result_json)
 
     def _on_error(self, bus, msg):
         self.error = msg.parse_error()
-        logger.error(self.error)
+        self.logger.error(self.error)
         self.finish_request()
         if self.error_handler:
             self.error_handler(self.error[0].message)
 
     def _on_eos(self, bus, msg):
-        logger.info('%s: Pipeline received eos signal' % self.request_id)
+        self.logger.info('%s: Pipeline received eos signal' % self.request_id)
         self.finish_request()
         if self.eos_handler:
             self.eos_handler[0](self.eos_handler[1])
@@ -154,7 +154,7 @@ class DecoderPipeline2(object):
         return self.asr.set_property("adaptation-state", adaptation_state)
 
     def finish_request(self):
-        logger.info("%s: Resetting decoder state" % self.request_id)
+        self.logger.info("%s: Resetting decoder state" % self.request_id)
         if self.output_dir:
             self.filesink.set_state(Gst.State.NULL)
             self.filesink.set_property('location', "/dev/null")
@@ -164,9 +164,9 @@ class DecoderPipeline2(object):
 
     def init_request(self, id, caps_str):
         self.request_id = id
-        logger.info("%s: Initializing request" % (self.request_id))
+        self.logger.info("%s: Initializing request" % (self.request_id))
         if caps_str and len(caps_str) > 0:
-            logger.info("%s: Setting caps to %s"
+            self.logger.info("%s: Setting caps to %s"
                         % (self.request_id, caps_str))
             caps = Gst.caps_from_string(caps_str)
             self.appsrc.set_property("caps", caps)
@@ -188,15 +188,15 @@ class DecoderPipeline2(object):
         self.set_adaptation_state("")
 
     def process_data(self, data):
-        logger.debug('%s: Pushing buffer of size %d to pipeline'
+        self.logger.debug('%s: Pushing buffer of size %d to pipeline'
                      % (self.request_id, len(data)))
         buf = Gst.Buffer.new_allocate(None, len(data), None)
         buf.fill(0, data)
         self.appsrc.emit("push-buffer", buf)
-        logger.debug('%s: Pushing buffer done' % self.request_id)
+        self.logger.debug('%s: Pushing buffer done' % self.request_id)
 
     def end_request(self):
-        logger.info("%s: Pushing EOS to pipeline" % self.request_id)
+        self.logger.info("%s: Pushing EOS to pipeline" % self.request_id)
         self.appsrc.emit("end-of-stream")
 
     def set_result_handler(self, handler):
@@ -212,7 +212,7 @@ class DecoderPipeline2(object):
         self.error_handler = handler
 
     def cancel(self):
-        logger.info("%s: Sending EOS to pipeline in order to cancel\
+        self.logger.info("%s: Sending EOS to pipeline in order to cancel\
  processing" % self.request_id)
         self.appsrc.emit("end-of-stream")
-        logger.info("%s: Cancelled pipeline" % self.request_id)
+        self.logger.info("%s: Cancelled pipeline" % self.request_id)
