+++ test/unittests/test-buffer.py
@@ -16,7 +16,7 @@ appsrc.link(filesink)
 pipeline.set_state(Gst.State.PLAYING)
 
 data = "1234" * 12
-print "Using data: %s" % data
+print("Using data: {}".format(data))
 
 buf = Gst.Buffer.new_allocate(None, len(data), None)
 buf.fill(0, data)
@@ -28,4 +28,4 @@ pipeline.send_event(Gst.Event.new_eos())
 
 result = open("test.dat").read()
 
-print "Result    : %s" % result
+print("Result    : {}".format(result))
